# server/api/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: StuHelper API
  version: 1.0.0
  description: StuHelper 学习助手平台 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080
    description: Development
  - url: https://api.stuhelper.com
    description: Production
...
paths:
...
  /api/v1/admin/student-verifications/{userID}:
    put:
      tags:
        - Admin
      operationId: reviewStudentVerification
      summary: 审核学生认证
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: userID
          in: path
          required: true
          description: 用户 ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewStudentVerificationRequest'
      responses:
        '200':
          description: 审核成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/admin/school-configs/{schoolID}:
    put:
      tags:
        - Admin
      operationId: updateSchoolConfig
      summary: 更新学校配置
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: schoolID
          in: path
          required: true
          description: 学校 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSchoolConfigRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/system-configs:
    get:
      tags:
        - Admin
      operationId: listSystemConfigs
      summary: 获取系统配置列表
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: 系统配置列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SystemConfig'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/system-configs/{key}:
    put:
      tags:
        - Admin
      operationId: updateSystemConfig
      summary: 更新系统配置
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          description: 配置键名
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSystemConfigRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required:
                      - data
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageData'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '500':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseBody'
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          enum:
            - true
    ErrorResponseBody:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: A00400
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    MessageData:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    ReviewStudentVerificationRequest:
      type: object
      required:
        - approved
      properties:
        approved:
          type: boolean
        rejectionReason:
          description: 驳回理由，可留空
          type:
            - string
            - 'null'
...
    UpdateSystemConfigRequest:
      type: object
      required:
        - value
      properties:
        value:
          type: string
...
    UpdateSchoolConfigRequest:
      type: object
      properties:
        schoolName:
          type: string
          maxLength: 100
        verificationMethod:
          type: string
          enum:
            - ldap
            - manual
        ldapConfig:
          $ref: '#/components/schemas/SchoolLDAPConfigInput'
        academicDbTable:
          type: string
          maxLength: 100
        consentText:
          type: string
        manualFormFields:
          type: array
          items:
            $ref: '#/components/schemas/ManualFieldDescriptor'
        enabled:
          type: boolean
...
    ManualFieldDescriptor:
      type: object
      required:
        - key
        - label
        - type
        - required
      properties:
        key:
          type: string
          description: 动态字段唯一标识
        label:
          type: string
          description: 字段展示文案
        type:
          type: string
          enum:
            - text
            - textarea
            - select
            - date
        required:
          type: boolean
        options:
          type:
            - array
            - 'null'
          items:
            type: string
          description: select 类型可选项
        placeholder:
          type:
            - string
            - 'null'
          description: 输入占位文案
...
    SchoolLDAPConfigInput:
      type: object
      properties:
        url:
          type: string
        baseDN:
          type: string
        systemBindDN:
          type: string
        systemBindPassword:
          type: string
        useTLS:
          type: boolean
        insecureSkipVerify:
          type: boolean
    SystemConfig:
      type: object
      required:
        - key
        - value
        - updatedAt
      properties:
        key:
          type: string
        value:
          type: string
        description:
          type:
            - string
            - 'null'
        updatedAt:
          type: string
          format: date-time
...